Move functions to process feeds to dedicated file
